Finally, add the videos from EuroLLVM 2014.
Unfortunately, there were some unforseen problems with the recording, which was
done by the University's Media Services division. This means that in some cases
we only have audio and video of the speaker but no recording of the screen, and
5 out of the 18 presentations were not recorded at all due to equipment failure.
We have expressed our strong dissatisfaction with this poor service, but
unfortunately there is nothing more that can be done at this stage.

The videos are on YouTube to faciliate viewing on different devices, but the
llvm.org server also holds an archive copy of the raw video files.

llvm-svn: 210319