Major update to relicensing page.
Summary:
Updates the relicensing page with all the details of the individual agreement
process and a major update to the remaining items and planned schedule.

Also updates the list of companies to include those we've already contacted to
avoid large numbers of people repeatedly pinging them when they are already
well and truly looped into the process.

Reviewers: hfinkel

Reviewed By: hfinkel

Subscribers: hfinkel, sanjoy, mcrosier, arphaman, llvm-commits

Differential Revision: https://reviews.llvm.org/D53269

llvm-svn: 344656