[LP-0004] Proposal for project governance (#54)
This PR introduces a proposal for project governance for the LLVM project. The proposal has gone through numerous iterations over the last year.

Earlier this week the review managers met to provide additional feedback. The notes from that meeting are posted to Discourse: https://discourse.llvm.org/t/rfc-llvm-project-governance-october-2024-update/82559

One of the actions from that meeting was to review and approve merging this PR without additional changes. A new PR will be posted shortly to address the specific feedback from the review manager meeting.

Thank you everyone for your feedback, contributions and support as we've worked through this proposal.

Co-authored-by: Louis Dionne <ldionne.2@gmail.com>