Remove non-GSoC "open projects"
These lists of "open projects" are badly outdated, and nobody is
maintaining them. And I don't think anyone wants to maintain them
in this place and format. A very large fraction of the "open"
projects have already been implemented.

I think it would be best to just delete these entirely.